Remove the Call/Return instructions.
These are not necessary since they can be implemented in just two
primitive instructions:

    push:
        STI PC, SP
        ADD PC, 1, PC

    pop:
        ADD PC, -1, PC
        LOI PC, SP

The stack pointer register itself may be eliminated in a subsequent
iteration, since it no longer has any special meaning. Will decide later
whether to rename it to "R5" and keep it, or to eliminate it along with
"R4" to reduce to only 4 registers, and thus a smaller register selection
mux.