Add a note about the behavior of the condition bit.
The condition bit will be part of the input into the control state machine,
allowing some state transitions to vary based on this bit. However, each
conditional instruction is expected to have an unconditional counterpart
that is implemented by simply forcing the condition bit high using the
LSB of the opcode.